requirements.txt
requests>=2.2.1
lxml>=2.3.2
boto3>=1.3.0
PyRSS2Gen==1.1
python-dateutil>=2.3
